requirements.txt
streamlit 
st_img_carousel.img_carousel